pyproject.toml
[tool.poetry]
name = "bristol"
version = "0.1.0"
description = ""
authors = ["Komron Khisomov <normok_9595@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
allure-pytest = "^2.13.2"
selenium = "^4.9.1"
loguru = "^0.7.0"
psycopg2-binary = "^2.9.6"
curlify = "^2.2.1"
webdriver-manager = "^3.8.6"
beautifulsoup4 = "^4.12.2"
testit-importer-allure = "^1.3.3"
flake8 = "^6.1.0"
black = "^23.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
